chore: layout refactor
