+++ web/components/headers/index.ts
@@ -18,3 +18,4 @@ export * from "./project-draft-issues";
 export * from "./project-archived-issue-details";
 export * from "./project-archived-issues";
 export * from "./project-issue-details";
+export * from "./user-profile";
