--- web/layouts/app-layout-legacy/layout.tsx
@@ -1,1 +1,0 @@
-export const AppLayoutLegacy = () => <></>;
