--- web/layouts/auth-layout-legacy/index.ts
@@ -1,3 +1,0 @@
-export * from "./project-authorization-wrapper";
-export * from "./workspace-authorization-wrapper";
-export * from "./user-authorization-wrapper";
