+++ web/layouts/profile-layout/layout.tsx
@@ -1,32 +1,22 @@
-import { useRouter } from "next/router";
 // hooks
 import { useWorkspaceMyMembership } from "contexts/workspace-member.context";
 // layouts
-import { WorkspaceAuthorizationLayout } from "layouts/auth-layout-legacy";
+import { AppLayout } from "layouts/app-layout";
 // components
 import { ProfileNavbar, ProfileSidebar } from "components/profile";
 // ui
-import { Breadcrumbs, BreadcrumbItem } from "@plane/ui";
+import { UserProfileHeader } from "components/headers";
 
 type Props = {
   children: React.ReactNode;
   className?: string;
 };
 
-export const ProfileAuthWrapper = (props: Props) => {
-  const router = useRouter();
-  return (
-    <WorkspaceAuthorizationLayout
-      breadcrumbs={
-        <Breadcrumbs onBack={() => router.back()}>
-          <BreadcrumbItem title="User Profile" />
-        </Breadcrumbs>
-      }
-    >
-      <ProfileLayout {...props} />
-    </WorkspaceAuthorizationLayout>
-  );
-};
+export const ProfileAuthWrapper = (props: Props) => (
+  <AppLayout header={<UserProfileHeader />}>
+    <ProfileLayout {...props} />
+  </AppLayout>
+);
 
 const ProfileLayout: React.FC<Props> = ({ children, className }) => {
   const { memberRole } = useWorkspaceMyMembership();
