+++ web/layouts/setting-layout/project-setting-layout/layout.tsx
@@ -1,8 +1,7 @@
 import { FC, ReactNode } from "react";
 // layouts
-import { UserAuthWrapper, ProjectAuthWrapper, WorkspaceAuthWrapper } from "layouts/auth-layout";
+import { AppLayout } from "layouts/app-layout";
 // components
-import { AppSidebar } from "layouts/app-layout";
 import { ProjectSettingsSidebar } from "./sidebar";
 
 export interface IProjectSettingLayout {
@@ -15,26 +14,14 @@ export const ProjectSettingLayout: FC<IProjectSettingLayout> = (props) => {
 
   return (
     <>
-      <UserAuthWrapper>
-        <WorkspaceAuthWrapper>
-          <ProjectAuthWrapper>
-            <div className="relative flex h-screen w-full overflow-hidden">
-              <AppSidebar />
-              <main className="relative flex flex-col h-full w-full overflow-hidden bg-custom-background-100">
-                {header}
-                <div className="h-full w-full overflow-hidden">
-                  <div className="flex gap-2 relative h-full w-full overflow-x-hidden overflow-y-scroll">
-                    <div className="w-80 pt-8 overflow-y-hidden flex-shrink-0">
-                      <ProjectSettingsSidebar />
-                    </div>
-                    {children}
-                  </div>
-                </div>
-              </main>
-            </div>
-          </ProjectAuthWrapper>
-        </WorkspaceAuthWrapper>
-      </UserAuthWrapper>
+      <AppLayout header={header} withProjectWrapper>
+        <div className="flex gap-2 h-full w-full overflow-x-hidden overflow-y-scroll">
+          <div className="w-80 pt-8 overflow-y-hidden flex-shrink-0">
+            <ProjectSettingsSidebar />
+          </div>
+          {children}
+        </div>
+      </AppLayout>
     </>
   );
 };
