+++ web/pages/[workspaceSlug]/profile/[userId]/subscribed.tsx
@@ -3,7 +3,7 @@ import React from "react";
 // contexts
 import { ProfileIssuesContextProvider } from "contexts/profile-issues-context";
 // layouts
-import { ProfileAuthWrapper } from "layouts/profile-layout";
+import { ProfileAuthWrapper } from "layouts/profile-layout/layout";
 // components
 import { ProfileIssuesView } from "components/profile";
 // types
